requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   httpx
    #   openai
black==23.12.1
    # via -r requirements.in
build==1.0.3
    # via pip-tools
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
cfgv==3.4.0
    # via pre-commit
click==8.1.7
    # via
    #   black
    #   pip-tools
coverage[toml]==7.4.0
    # via
    #   coverage
    #   pytest-cov
distlib==0.3.8
    # via virtualenv
distro==1.9.0
    # via openai
filelock==3.13.1
    # via virtualenv
flake8==6.1.0
    # via -r requirements.in
h11==0.14.0
    # via httpcore
httpcore==1.0.4
    # via httpx
httpx==0.27.0
    # via openai
identify==2.5.33
    # via pre-commit
idna==3.6
    # via
    #   anyio
    #   httpx
iniconfig==2.0.0
    # via pytest
isort==5.13.2
    # via -r requirements.in
mccabe==0.7.0
    # via flake8
mypy==1.8.0
    # via -r requirements.in
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
nodeenv==1.8.0
    # via pre-commit
openai==1.12.0
    # via -r requirements.in
packaging==23.2
    # via
    #   black
    #   build
    #   pytest
pathspec==0.12.1
    # via black
pip-tools==7.3.0
    # via -r requirements.in
platformdirs==4.1.0
    # via
    #   black
    #   virtualenv
pluggy==1.3.0
    # via pytest
pre-commit==3.6.0
    # via -r requirements.in
pycodestyle==2.11.1
    # via flake8
pydantic==2.6.2
    # via openai
pydantic-core==2.16.3
    # via pydantic
pyflakes==3.1.0
    # via flake8
pyproject-hooks==1.0.0
    # via build
pytest==7.4.4
    # via
    #   -r requirements.in
    #   pytest-cov
pytest-cov==4.1.0
    # via -r requirements.in
python-dotenv==1.0.1
    # via -r requirements.in
pyyaml==6.0.1
    # via pre-commit
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
tqdm==4.66.2
    # via openai
typing-extensions==4.9.0
    # via
    #   mypy
    #   openai
    #   pydantic
    #   pydantic-core
virtualenv==20.25.0
    # via pre-commit
wheel==0.42.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
